# api/openapi.yaml
openapi: 3.1.0
info:
  title: StackPilot API
  version: 0.0.0
  description: StackPilot local control-plane API contract.
servers:
  - url: /
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /api/v1/auth/bootstrap:
    post:
      operationId: issueBrowserBootstrap
      summary: Issue a short-lived single-use browser bootstrap code
      security:
        - bearerAuth: []
      responses:
        "201":
          description: One-time bootstrap code issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/auth/session:
    post:
      operationId: exchangeBrowserSession
      summary: Exchange a same-origin fragment bootstrap for a browser session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionExchangeRequest"
      responses:
        "201":
          description: Browser session created and set as an HttpOnly cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      operationId: renewBrowserSession
      summary: Renew an active browser session and rotate its CSRF value
      description: Advances the 30-minute renewal expiry without exceeding the eight-hour absolute limit, rotates CSRF, and refreshes the HttpOnly cookie. This state-changing recovery GET is same-origin credentialed and is never cacheable.
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Session was renewed and its CSRF value and cookie were rotated.
          headers:
            Set-Cookie:
              description: Refreshed HttpOnly session cookie with the authoritative renewal expiry.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: revokeBrowserSession
      summary: Revoke the current browser session
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "204":
          description: Browser session revoked and cookie expired.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/auth/token/rotate:
    post:
      operationId: rotateLocalAuthenticationToken
      summary: Rotate the OS-protected local token and invalidate browser sessions
      parameters:
        - $ref: "#/components/parameters/CSRF"
      responses:
        "200":
          description: Local token rotated; browser callers receive an expired session cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRotationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/incidents:
    get:
      operationId: listIncidents
      summary: List bounded diagnostic incidents for one workspace
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/WorkspaceId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Newest-first incident summaries with redacted bounded context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentList"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/incidents/{incidentId}:
    get:
      operationId: getIncident
      summary: Read one incident and all versioned deterministic analyses
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
            pattern: "^inc_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
      responses:
        "200":
          description: Traceable incident evidence and diagnoses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/incidents/{incidentId}/analyze:
    post:
      operationId: analyzeIncident
      summary: Queue a read-only evidence refresh and deterministic rule analysis
      parameters:
        - name: incidentId
          in: path
          required: true
          schema:
            type: string
            pattern: "^inc_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            maxLength: 128
        - $ref: "#/components/parameters/CSRF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "202":
          description: The read-only analyze Operation was queued or replayed idempotently.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Long-lived local token read only from OS-protected storage by CLI clients.
    sessionAuth:
      type: apiKey
      in: cookie
      name: stackpilot_session
  parameters:
    CSRF:
      name: X-StackPilot-CSRF
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - AUTH_BOOTSTRAP_INVALID
        - EVENT_CURSOR_EXPIRED
        - FEATURE_NOT_ENABLED
        - AUTH_BROWSER_REQUEST_REJECTED
        - AUTH_CAPACITY_REACHED
        - AUTH_SESSION_INVALID
        - AUTH_TOKEN_INVALID
        - CONTROL_PLANE_RESTARTED
        - COMPOSE_CONFIG_INVALID
        - COMPOSE_BUILD_CONFIG_INVALID
        - COMPOSE_BUILD_FAILED
        - COMPOSE_BUILD_TIMEOUT
        - COMPOSE_DISCOVERY_FAILED
        - COMPOSE_INSPECT_FAILED
        - COMPOSE_LOG_FOLLOW_FAILED
        - COMPOSE_LIFECYCLE_INVALID
        - COMPOSE_LIFECYCLE_TIMEOUT
        - COMPOSE_NOT_FOUND
        - COMPOSE_OVERRIDE_CONFLICT
        - COMPOSE_OVERRIDE_INVALID
        - COMPOSE_PREFLIGHT_TIMEOUT
        - COMPOSE_PROJECT_IDENTITY_MISMATCH
        - COMPOSE_PROJECT_NOT_FOUND
        - COMPOSE_START_FAILED
        - COMPOSE_STOP_FAILED
        - CONTAINER_UNHEALTHY
        - COMPOSE_VERSION_UNSUPPORTED
        - DOCKER_DAEMON_UNAVAILABLE
        - DOCKER_NOT_FOUND
        - DOCKER_VERSION_UNSUPPORTED
        - HEALTH_NOT_READY
        - HEALTH_READINESS_TIMEOUT
        - HTTP_BODY_MISMATCH
        - HTTP_STATUS_MISMATCH
        - HTTP_TIMEOUT
        - INTERNAL_ERROR
        - LOG_CURSOR_EXPIRED
        - IDEMPOTENCY_KEY_REUSED
        - MANIFEST_DUPLICATE_KEY
        - MANIFEST_CYCLE_DETECTED
        - MANIFEST_DURATION_INVALID
        - MANIFEST_FILE_TOO_LARGE
        - MANIFEST_HEALTH_TARGET_UNSAFE
        - MANIFEST_CHANGED
        - MANIFEST_MULTIPLE_DOCUMENTS
        - MANIFEST_NOT_REGULAR_FILE
        - MANIFEST_PATH_OUTSIDE_WORKSPACE
        - MANIFEST_PORT_RANGE_INVALID
        - MANIFEST_REFERENCE_NOT_FOUND
        - MANIFEST_SCHEMA_INVALID
        - MANIFEST_SEMANTIC_INVALID
        - MANIFEST_TEMPLATE_INVALID
        - MANIFEST_UNKNOWN_FIELD
        - MANIFEST_YAML_INVALID
        - METHOD_NOT_ALLOWED
        - OPERATION_ALREADY_ACTIVE
        - OPERATION_INVALID_STATE
        - OPERATION_NOT_CANCELLABLE
        - OPERATION_NOT_FOUND
        - PLATFORM_NOT_SUPPORTED
        - PORT_CONFLICT
        - PROCESS_EXITED
        - PROCESS_IDENTITY_MISMATCH
        - PROCESS_START_FAILED
        - PROCESS_STOP_FAILED
        - REQUEST_VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - RUNNER_NOT_FOUND
        - RUNNER_PATH_UNSAFE
        - RUNNER_VERSION_CHECK_FAILED
        - RUNNER_VERSION_CHECK_TIMEOUT
        - SECRET_INVALID
        - SECRET_NOT_FOUND
        - SECRET_VERSION_CONFLICT
        - TCP_REFUSED
        - TCP_TIMEOUT
        - SUPERVISOR_EXITED
        - SUPERVISOR_UNAVAILABLE
        - WORKSPACE_ALREADY_REGISTERED
        - WORKSPACE_ID_REQUIRED
        - WORKSPACE_MANIFEST_UNAVAILABLE
        - WORKSPACE_NOT_FOUND
        - WORKSPACE_PATH_INVALID
        - WORKSPACE_SCRIPT_NOT_FOUND
        - WORKSPACE_SCRIPT_OUTSIDE
        - WORKSPACE_SCRIPT_TYPE_UNSUPPORTED
        - WORKSPACE_SCRIPT_ENCODING_UNSUPPORTED
        - WORKSPACE_SCRIPT_TOO_LARGE
        - WORKSPACE_SCRIPT_SYNTAX_UNSUPPORTED
        - WORKSPACE_SCRIPT_DANGEROUS
        - WORKSPACE_SCRIPT_REFERENCE_CYCLE
        - WORKSPACE_IMPORT_INCOMPLETE
        - WORKSPACE_IMPORT_PORT_UNCONFIRMED
        - WORKSPACE_IMPORT_DEPENDENCY_UNCONFIRMED
        - WORKSPACE_IMPORT_SOURCE_CHANGED
        - WORKSPACE_IMPORT_DRAFT_EXPIRED
        - WORKSPACE_MANIFEST_CONFLICT
        - WORKSPACE_MANIFEST_WRITE_FAILED
        - WORKSPACE_EDIT_RUNTIME_ACTIVE
        - WORKSPACE_UNREGISTER_RUNTIME_ACTIVE
        - WORKSPACE_RELINK_SYSTEM_MISMATCH
        - WORKSPACE_SYSTEM_ID_CHANGED
    TokenRotationResponse:
      type: object
      additionalProperties: false
      required: [createdAt]
      properties:
        createdAt:
          type: string
          format: date-time
...
    BootstrapResponse:
      type: object
      additionalProperties: false
      required: [code, expiresAt]
      properties:
        code:
          type: string
          minLength: 43
          maxLength: 43
        expiresAt:
          type: string
          format: date-time
    SessionExchangeRequest:
      type: object
      additionalProperties: false
      required: [bootstrap]
      properties:
        bootstrap:
          type: string
          minLength: 43
          maxLength: 43
    SessionResponse:
      type: object
      additionalProperties: false
      required: [csrf, expiresAt]
      properties:
        csrf:
          type: string
          minLength: 43
          maxLength: 43
        expiresAt:
          type: string
          format: date-time
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
        - traceId
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        operationId:
          type:
            - string
            - "null"
          pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        traceId:
          type: string
          pattern: "^tr_[0-9a-f]{32}$"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    OperationId:
      type: string
      pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
...
    OperationRef:
      type: object
      additionalProperties: false
      required: [operationId, state]
      properties:
        operationId:
          $ref: "#/components/schemas/OperationId"
        state:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
...
    IncidentEvidence:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          type: string
          enum: [event, health, log]
        eventId:
          type: integer
          minimum: 1
        healthResultId:
          type: integer
          minimum: 1
        serviceInstanceId:
          type: string
          pattern: "^svi_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        logSequence:
          type: integer
          minimum: 1
    IncidentContext:
      type: object
      additionalProperties: false
      required: [schemaVersion, workspaceId, kind, triggerCode, windowStart, windowEnd, dependencies, ports, evidence, logs]
      properties:
        schemaVersion:
          type: string
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        systemInstanceId:
          type: string
        serviceInstanceId:
          type: string
        serviceId:
          $ref: "#/components/schemas/ManifestId"
        kind:
          type: string
          enum: [port-conflict, process-exit, readiness-timeout, liveness-failure, restart-limit, identity-mismatch, known-log-error]
        triggerCode:
          type: string
        windowStart:
          type: string
          format: date-time
        windowEnd:
          type: string
          format: date-time
        dependencies:
          type: object
          additionalProperties:
            type: string
        ports:
          type: object
          additionalProperties:
            type: integer
        evidence:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/IncidentEvidence"
        logs:
          type: array
          maxItems: 500
          items:
            type: object
            additionalProperties: true
    Incident:
      type: object
      additionalProperties: false
      required: [id, workspaceId, kind, severity, state, occurrenceCount, context, firstSeenAt, lastSeenAt]
      properties:
        id:
          type: string
          pattern: "^inc_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        workspaceId:
          $ref: "#/components/schemas/WorkspaceId"
        systemInstanceId:
          type: string
        serviceInstanceId:
          type: string
        serviceId:
          $ref: "#/components/schemas/ManifestId"
        kind:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        state:
          type: string
          enum: [open, resolved]
        occurrenceCount:
          type: integer
          minimum: 1
        context:
          $ref: "#/components/schemas/IncidentContext"
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    IncidentList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/Incident"
    IncidentAnalysis:
      type: object
      additionalProperties: false
      required: [id, engine, schemaVersion, result, createdAt]
      properties:
        id:
          type: integer
        engine:
          type: string
        schemaVersion:
          type: string
        result:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    IncidentDetail:
      type: object
      additionalProperties: false
      required: [incident, analyses]
      properties:
        incident:
          $ref: "#/components/schemas/Incident"
        analyses:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/IncidentAnalysis"
...
    ManifestId:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
...
    WorkspaceId:
      type: string
      pattern: "^ws_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
...
  responses:
    Unauthorized:
      description: Local token or browser session authentication failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Browser origin or CSRF validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServiceUnavailable:
      description: Authentication capacity is temporarily exhausted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BadRequest:
      description: The request or manifest syntax is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: The request conflicts with current catalog state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...